requirements.txt
asgiref
astroid
autopep8
Babel
black
cairocffi
CairoSVG
cffi
codecov
coverage
cssselect2
defusedxml
django-crispy-forms
crispy-bootstrap4
django-phonenumber-field
django-weasyprint
flake8
freezegun
gunicorn
html5lib
isort
lazy-object-proxy
mccabe
pep8
phonenumbers
Pillow
pycodestyle
pycparser
pylint
PyPDF2
Pyphen
python-dateutil
pytz
reportlab
selenium
six
sqlparse
tinycss2
typed-ast
urllib3
WeasyPrint
webencodings
wrapt
xhtml2pdf
psycopg2
